Issue 14: Purge stale facts file
This module now defaults to using /opt/puppetlabs/facter/facts.d/ for
its external facts file for Facter >= 2.4.2 and
/etc/puppetlabs/facter/facts.d/ otherwise. When updating the module from
a previous version, agents using the new path may still have the facts
in the /etc/ path, which seems to take precedence. Therefore, the
version isn't bumped in that facter file and Bamboo thinks it needs to
update each run.

Fixes #14